Add project skeleton with js and php routers
This is a project skeleton that uses local, foreground nginx and php-fpm
to send queries having something.php in first part of query path to index.php,
anything that ends with .js to bundle.js (generated with webpack) and the rest —
to index.html which is a piece of HTML where

To run nginx, navigate to the project root (.../file-server),
if your system supports it, run nix-shell,
tweak document root in nginx.conf to point to the output of $(pwd)
and then run it the following way:
``nginx -c $(pwd)/nginx.conf -p $(pwd)``

To run php-fpm, navigate to the projecct root,
if your system supports it, run nix-shell,
then run ``php-fpm -y ./fpm.conf -p $(pwd)``.

Now you can navigate to http://localhost:8074/anything.php to get
the following JSON:
``{"error": "Nothing is implemented. Nothing, Carl."}``.

When you navigate to an arbitrary URL on this server, say,
you'll get the following result:

```
**Controller: foo**

Args:
  bar
  baz=42
```

Now we can build a JS-router for different "controllers" and PHP-handlers for
data manipulation and serving it trough JSONRPC 2.0 API.