Add routing for JS and skeleton of application
We use promises to check with the server API for permissions
to access certain parts of infrastructure, and — depending on that —
route user who comes from a certain url through initial component
to be rendered.

From there it's a single-page application. In order for user to share
a certain page with his peers, he or she will have to use "Link here"
component (to be written) instead of document.location.